+++ livelossplot/outputs/neptune.py
@@ -1,7 +1,7 @@
 import neptune
 
 from livelossplot.main_logger import MainLogger
-from livelossplot.output_plugins.base_output import BaseOutput
+from livelossplot.outputs.base_output import BaseOutput
 
 
 class NeptuneLogger(BaseOutput):